Configure bot to only respond when chat is 'active'
Chat will be considered active if any message has been received from
a user within the last 6 minutes. If there has been a message, a
quote is sent, and the 1 hour timer starts counting down.
If there has not been a message, only a 5 minute back off will be applied.